// client/src/app/employee.ts
export class Employee{
  _id: string;
  first_name: string;
  last_name: string;
  birth_date: string;
  pay_rate: string;
  bonus: string;
  reimbursement: string;
  retirement: string;
  stateSelected: string;
}

export class StateTax{
  _id: string;
  name: string;
  abbreviation: string;
  tax: string;
}